requirements.txt
torch==2.3.0
torchvision==0.18.0
diffusers==0.29.0
accelerate==0.30.1
peft==0.11.1
transformers==4.41.2
datasets==2.19.1
safetensors==0.4.3
pillow==10.3.0
tqdm==4.66.2
huggingface-hub==0.23.0
requests==2.32.3
beautifulsoup4==4.12.3
tqdm==4.66.2
aiohttp
async-timeout